moving search components into folders
