+++ app/javascript/components/HomePageContent.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import SearchPanel from './SearchPanel'
-import ResultsPanel from './ResultsPanel'
+import SearchPanel from 'components/search/controls/SearchPanel'
+import ResultsPanel from 'components/search/results/ResultsPanel'
 import StudySearchProvider from 'providers/StudySearchProvider'
 import SearchFacetProvider from 'providers/SearchFacetProvider'
 import UserProvider from 'providers/UserProvider'
