+++ app/javascript/components/search/controls/FacetControl.js
