+++ app/javascript/components/search/controls/FacetsAccordion.js
