+++ app/javascript/components/search/controls/FacetsPanel.js
