+++ app/javascript/components/search/controls/FilterSlider.js
