+++ app/javascript/components/search/controls/FiltersBoxSearchable.js
