+++ app/javascript/components/search/controls/FiltersSearchBar.js
