+++ app/javascript/components/search/controls/KeywordSearch.js
