+++ app/javascript/components/search/controls/MoreFacetsButton.js
