+++ app/javascript/components/search/controls/SearchPanel.js
