+++ app/javascript/components/search/controls/slider/components.js
