+++ app/javascript/components/search/results/PagingControl.js
