+++ app/javascript/components/search/results/ResultsPanel.js
