+++ app/javascript/components/search/results/SearchQueryDisplay.js
