+++ app/javascript/components/search/results/Study.js
