+++ app/javascript/components/search/results/StudyResultsContainer.js
