+++ test/js/search/combined-search.test.js
@@ -5,8 +5,8 @@ import * as Reach from '@reach/router'
 
 const fetch = require('node-fetch');
 
-import FacetControl from 'components/FacetControl';
-import KeywordSearch from 'components/KeywordSearch';
+import FacetControl from 'components/search/controls/FacetControl';
+import KeywordSearch from 'components/search/controls/KeywordSearch';
 import { PropsStudySearchProvider } from 'providers/StudySearchProvider';
 import * as ScpAPI from 'lib/scp-api'
