+++ test/js/search/filter-slider.test.js
@@ -5,7 +5,7 @@ import * as Reach from '@reach/router'
 
 const fetch = require('node-fetch')
 
-import FacetControl from 'components/FacetControl'
+import FacetControl from 'components/search/controls/FacetControl'
 import { PropsStudySearchProvider } from 'providers/StudySearchProvider'
 import { SearchSelectionContext } from 'providers/SearchSelectionProvider'
