+++ test/js/search/more-facets.test.js
@@ -5,7 +5,7 @@ import * as Reach from '@reach/router'
 
 const fetch = require('node-fetch');
 
-import MoreFacetsButton from 'components/MoreFacetsButton';
+import MoreFacetsButton from 'components/search/controls/MoreFacetsButton';
 import { PropsStudySearchProvider } from 'providers/StudySearchProvider';
 
 import * as ScpAPI from 'lib/scp-api'
