+++ test/js/search/search-query-display.test.js
@@ -3,10 +3,10 @@ import { mount } from 'enzyme'
 
 const fetch = require('node-fetch')
 
-import SearchQueryDisplay, { ClearAllButton } from 'components/SearchQueryDisplay'
+import SearchQueryDisplay, { ClearAllButton } from 'components/search/results/SearchQueryDisplay'
 import { PropsStudySearchProvider } from 'providers/StudySearchProvider'
-import KeywordSearch from 'components/KeywordSearch'
-import FacetControl from 'components/FacetControl'
+import KeywordSearch from 'components/search/controls/KeywordSearch'
+import FacetControl from 'components/search/controls/FacetControl'
 
 
 const oneStringFacet = [
